Server/requirements.txt
docopt==0.4.0
Flask==0.10.1
Flask-Cors==2.0.1
httplib2==0.9.1
influxdb==2.3.0
itsdangerous==0.24
Jinja2==2.7.3
logging==0.4.9.6
mandrill==1.0.57
MarkupSafe==0.23
pyserial==2.7
python-forecastio==1.3.2
pytz==2015.4
requests==2.7.0
schedule==0.3.1
six==1.9.0
twilio==4.4.0
Werkzeug==0.10.4
